+++ frontend/src/App.jsx
@@ -1,13 +1,7 @@
 import React from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import ProtectedRoute from "./components/common/ProtectedRoute";
-// import Navbar from "./components/navbar/navbar";
 import HomePage from "./pages/Home/homepage"; //Landing page Venta
-// import Resgistrarse from "./components/resgistrarse";
-// import IniciarSesion from "./components/iniciarSesion";
-// import SobreNosotros from "./pages/sobreNosotros"; // ruta quienes somos
-// import PoliticasPrivacidad from "./components/politicasPrivacidad"; // ruta politica de privacidad
-// import TerminosCond from "./components/terminosCond"; // ruta terminos y condiciones
 
 //Importaciones de Luis (Autenticacion)
 import Login from "./pages/Auth/Login";
@@ -17,7 +11,7 @@ import EmailRecovery from "./pages/Auth/EmailRecovery";
 import PasswordReset from "./pages/Auth/PasswordReset";
 import Terms from "./pages/Legal/Terms";
 import PrivacyPolicy from "./pages/Legal/PrivacyPolicy";
-import Simulation from "./pages/Simulation";
+import DashboardCreateProject from "./pages/DashboardCreateProject";
 import Error404 from "./pages/error404";
 import { Toaster } from 'react-hot-toast';
 import Dashboard from "./pages/Dashboard";
@@ -27,11 +21,7 @@ const App = () => {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Navigate to="/home" />} />
-        <Route path="/home" element={<HomePage />} />
-        {/* <Route path="/sobre-nosotros" element={<SobreNosotros />} />
-        <Route path="/iniciar-sesion" element={<IniciarSesion />} />
-        <Route path="/registrarse" element={<Resgistrarse />} />
-        <Route path="/politica-de-privacidad" element={<PoliticasPrivacidad />} />*/}       
+        <Route path="/home" element={<HomePage />} />     
         <Route path="/iniciar-sesion" element={<Login />} />
         <Route path="/registrarse" element={<Register />} />
         <Route path="/recuperar-contrasena" element={<PasswordRecovery />} />
@@ -39,9 +29,9 @@ const App = () => {
         <Route path="/password-reset/:token" element={<PasswordReset />} />
         <Route path="/terminos" element={<Terms />} />
         <Route path="/politica-de-privacidad" element={<PrivacyPolicy />} />
-        <Route path="/simulacion" element={
+        <Route path="/dashboard-create-project" element={
           <ProtectedRoute>
-            <Simulation />
+            <DashboardCreateProject />
           </ProtectedRoute>
         }/>
         <Route path="/dashboard" element={
