+++ frontend/src/components/legal/LegalPage.jsx
