--- frontend/src/components/sobreNosotros/header.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const header = () => {
-  return (
-    <div className='relative py-15 bg-gradient-to-b from-purple-900 to-purple-700 text-white'>  
-      
-    </div>
-  )
-}
-
-export default header
