+++ frontend/src/pages/Legal/sobreNosotros.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const SobreNosotros = () => {
+    return (
+      <div className="text-center mt-10 text-white">
+        <h2 className="text-4xl font-bold">Sobre Nosotros</h2>
+        <p className="mt-4 text-lg">Aquí va la información sobre tu microempresa o equipo.</p>
+      </div>
+    );
+  };
+  
+  export default SobreNosotros;
+
